Created market.py and analysis.py
market.py contains the class Market which is meant to act as a way to
quickly access a market's data.

analysis.py contains some functions that give a bit more insight, for
example giving the markets with the most volume or giving a set of
tuples representing a time series of prices.